test(api): share runtime-resolution guard via CommonJS for Jest
